--- .gitignore
@@ -1,19 +1,0 @@
-# System files
-.DS_Store
-
-# Object files
-*.o
-
-# Libraries
-*.lib
-*.a
-
-# Executables
-*.out
-*.exe
-
-# Temp files
-*.swp
-
-# Debug files
-*.dSYM/
