+++ Makefile
@@ -12,7 +12,10 @@ NAME = minishell
 SRCS = main.c tester.c syntax.c utils.c utils2.c utils3.c token_utils.c ft_quotes_split.c free_funcs.c type_check.c syntax2.c expand.c build_commands.c \
 	pedro_files/ft_pwd.c pedro_files/ft_echo.c\
 	pedro_files/ft_built_in.c pedro_files/ft_cd.c\
-	pedro_files/ft_env.c pedro_files/ft_built_errors.c
+	pedro_files/ft_env.c pedro_files/ft_built_errors.c \
+	pedro_files/ft_execute_m.c pedro_files/ft_execute.c \
+	pedro_files/utils.c 
+
 
 LIBFT = ./libraries/libft/libft.a
 PRINTF = ./libraries/printf/ft_printf.a
